# updates to Gaussian map for sliding window inference