@ericspod 2022-10-09
For what it's worth I have this code for defining a gaussian kernel of arbitrary size:

```python
def gaussian_kernel(*shape, device, dtype, sigma=2.0):
    dims = [torch.arange(-s // 2 + s % 2, s - s // 2).to(device, dtype) for s in shape]
    grid = torch.stack(torch.meshgrid(*dims, indexing="ij"))
    return (grid**2).sum(0).div_(-2 * sigma**2).exp_()
```

I think this is the standard definition though I don't think `gaussian_1d` produces an equivalent result.