@myron 2022-10-09
> For what it's worth I have this code for defining a gaussian kernel of arbitrary size:
> 
> ```python
> def gaussian_kernel(*shape, device, dtype, sigma=2.0):
>     dims = [torch.arange(-s // 2 + s % 2, s - s // 2).to(device, dtype) for s in shape]
>     grid = torch.stack(torch.meshgrid(*dims, indexing="ij"))
>     return (grid**2).sum(0).div_(-2 * sigma**2).exp_()
> ```
> 
> I think this is the standard definition though I don't think `gaussian_1d` produces an equivalent result.

yeah, your version looks good too. But it doesn't handle different sigma (per dimension).  With meshgrid you're allocating 3xDxWxH arrays (in case of 3D patch ), and compute exp over this 3D volume (DxWxH). In my version (separable) it only allocates DxWxH gpu mem, and computes torch.exp 3x(D+W+H) times, so it's a bit faster and takes less memory. Otherwise it's the same.