@ericspod 2022-10-09
I hadn't thought of different sigmas, but if the last line is `(grid**2).div_(-2 * sigma**2).sum(0).exp_()` then `sigma` can be a tensor of values instead to handle that case.